fix(r): use as.double for Measurement S3 method dispatch
as.numeric is a primitive alias for as.double in R, so S3 dispatch
looks for as.double.ClassName, not as.numeric.ClassName. Rename the
method and NAMESPACE registration to as.double.Measurement so that
as.numeric(measurement) works correctly.

+++ r/pragmastat/R/measurement.R
@@ -24,6 +24,6 @@ Measurement <- R6::R6Class(
 )
 
 #' @export
-as.numeric.Measurement <- function(x, ...) {
+as.double.Measurement <- function(x, ...) {
   x$value
 }
